requirements.txt
streamlit>=1.46.0
google-genai>=0.3.0
attrs>=21.0.0
pydantic>=2.0.0 